EAM-130 add base styles and variables
