+++ src/components/HomePage/AdvantagesList/AdvantageItem/AdvantageItem.module.scss
