+++ src/components/HomePage/AdvantagesList/AdvantageItem/AdvantageItem.tsx
