+++ src/components/HomePage/AdvantagesList/AdvantagesList.module.scss
@@ -1,6 +1,4 @@
 .advantagesList {
   display: flex;
   justify-content: space-between;
-  width: 1280px;
-  margin: 100px auto;
 }
