+++ src/components/HomePage/Hero/Hero.tsx
@@ -5,14 +5,14 @@ import classes from './Hero.module.scss';
 
 export const Hero: React.FC = () => {
   return (
-    <div className={classNames(classes.hero)}>
-      <div className={classNames(classes.hero_content)}>
+    <section className={classNames(classes.hero)}>
+      <div className={classNames(classes.hero_content, 'container')}>
         <h1 className={classNames(classes.hero_header)}>Магазин японських солодощів</h1>
         <p className={classNames(classes.hero_text)}>Унікальний вибір традиційних та сучасних смаколиків</p>
         <NavLink to="/catalog" className={classNames(classes.hero_btn)}>
           Біжу купляти
         </NavLink>
       </div>
-    </div>
+    </section>
   );
 };
