+++ src/components/HomePage/Presentation/PresentationItem/PresentationItem.module.scss
