+++ src/styles/helpers/vars.scss
@@ -34,4 +34,4 @@ $gl_radius: 0.8rem;
 $border: 0.1rem solid $border;
 
 //Transitions & Animation
-$transition: all 300ms cubic-bezier(0.53, 0.5, 0.53, 0.52) 0s;
+$transition: all 250ms cubic-bezier(0.53, 0.5, 0.53, 0.52) 0s;
